requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9

faiss-cpu==1.8.0.post1
numpy==1.26.4

sentence-transformers==3.0.1
transformers==4.43.4
torch==2.3.1

pydantic==2.9.2
